requirements.txt
PyQt6==6.6.1
PyQt6-Qt6==6.6.1
requests==2.31.0
Pillow==10.1.0
mss==9.0.1
opencv-python==4.8.1.78
pyttsx3==2.90
openai==1.3.7
python-dotenv==1.0.0
easyocr==1.7.0
ultralytics==8.0.206